refactor: fsd components
