+++ src/widgets/auth-actions/index.ts
@@ -1,1 +1,1 @@
-export { AuthActions } from "./ui";
+export * from "./ui";
