+++ src/widgets/complex-form/index.ts
@@ -1,1 +1,1 @@
-export { ComplexForm } from "./ui";
+export * from "./ui";
